Add WebOutputString IO. You can use it like that :
<calaos:output data="" data_type="" gui_type="string_out" id="output_1" name="Output String" request_type="POST" type="WebOutputString" url="http://192.168.1.72/cgi-bin/tts?voice=alice&amp;text=__##VALUE##__&amp;nocache=0" visible="true" />

url : is the url you want to contact and where __##VALUE##__ is the value you want to send. If you want to post data, use the data param. __##VALUE##__ is replaced each time the output change with the value set in calaos.
data : If you want to post data with your request, like json or xml for example, you can write the data here, and each occurence of __##VALUE##__ will be replaced with the current string in calaos.
data_type: it's MIME type of the data you want to send like application/json for example
request_type : It's the type of the request you want to send it can be one of the following : POST, PUT, GET